% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{mat_read}
\alias{mat_read}
\title{Writes to a MAT file}
\usage{
mat_read(matfile, ...)
}
\arguments{
\item{matfile}{a MAT file}

\item{...}{further arguments passe to \code{\link[R.matlab]{writeMat}}.}
}
\description{
This function is a soft-wrapper around \code{\link[R.matlab]{writeMat}}.
}
