% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{mat_trans}
\alias{mat_trans}
\title{Native Matlab translator for R objects}
\usage{
mat_trans(object, assign_name = TRUE, byrow = TRUE)
}
\arguments{
\item{object}{An R object. Accomodates translation for atomic vectors (numberic, character and logical),
and for matrix and data.frame.}

\item{assign_name}{whether to assign the name of the R object. Default is TRUE.}

\item{byrow}{whether to fill by row on by column. Onyl for atomic vectors.}
}
\description{
Translater R object into Matlab code. Can be used efficiently into a `mat_init` chain and passed into
`mat_run`.
}
